fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
pydantic==2.5.2
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.1
onnxruntime==1.17.0
transformers==4.49.0
numpy==1.26.4
onnx==1.17.0
